# api.yaml
swagger: '2.0'
info:
  title: Waiter Link API
  description: API to communicate between back end and front end of Waiter Link app
  version: 1.0.0
host: www.waiter.link
schemes:
  - https
basePath: /v1
produces:
  - application/json
paths:
  /api/signin:
    post:
      summary: Authenticate user with email and password
      description: |
        This is the one stop authentication to all the services of Waiter Link
         <br/>NEED REVISE
      tags:
        - Authentication
      parameters:
        - name: username
          in: query
          description: Username of user.
          required: true
          type: string
        - name: password
          in: query
          description: Password of user.
          required: true
          type: string
      responses:
        '200':
          description: Log in successfully
          schema:
            type: object
            properties:
              user:
                $ref: '#/definitions/User'
        '400':
          description: Error in log in
          schema:
            $ref: '#/definitions/Error'
...
  /api/waiterGetListWorkRestaurant:
    get:
      summary: Waiter get list of restaurant he/she is authenticated to work
      responses:
        '200':
          description: OK
          schema:
            type: object
            properties:
              restaurants:
                $ref: '#/definitions/Restaurant'
        '400':
          description: Error
          schema:
            $ref: '#/definitions/Error'
...
  /api/waiterStartShift:
    post:
      summary: Waiter start their shift and sign up to server a group of table
      parameters:
        - name: restaurant_id
          in: query
          description: Restaurant id number in database
          required: true
          type: integer
        - name: table_number
          in: query
          description: 'List of table number separated by comma, such as "2,3,4,7,8"'
          required: true
          type: string
      responses:
        '200':
          description: OK
          schema:
            type: object
            properties:
              shift_id:
                type: integer
                description: 'ID number of the shift this waiter just started'
        '400':
          description: Error
          schema:
            $ref: '#/definitions/Error'
...
  /api/waiterCloseShift:
    post:
      summary: Waiter retrieve list of shifts they are working
      parameters:
        - name: shift_id
          in: query
          description: Id of the shift to close
          required: true
          type: integer
      responses:
        '200':
          description: OK
        '400':
          description: Error
          schema:
            $ref: '#/definitions/Error'
  /api/waiterChangeServingTable:
    post:
      summary: Waiter change the tables he/she is serving
      parameters:
        - name: shift_id
          in: query
          description: Id of the shift to close
          required: true
          type: integer
        - name: table_number
          in: query
          description: 'List of table number separated by comma, such as "2,3,4,7,8"'
          required: true
          type: string
      responses:
        '200':
          description: OK
        '400':
          description: Error
          schema:
            $ref: '#/definitions/Error'
...
definitions:
  User:
    type: object
    properties:
      user_id:
        type: integer
        description: User id in the database
      manager_flag:
        type: boolean
        description: Flag to indicate if this user has manager privilege
      diner_flag:
        type: boolean
        description: Flag to indicate if this user has diner privilege
      waiter_flag:
        type: boolean
        description: Flag to indicate if this user has waiter privilege
...
  Restaurant:
    type: object
    properties:
      restaurant_id:
        type: integer
        description: Restaurant id in the database
      name:
        type: string
        description: Name of restaurant
      address:
        type: string
        description: Address of restaurant
      picture:
        type: string
        description: URL to logo/picture of the restaurant
      num_table:
        type: integer
        description: Number of table at this restaurant
...
  Error:
    type: object
    properties:
      code:
        type: integer
        format: int32
      message:
        type: string